[Back-fix] 노비 문서 랜덤 위치 뿌리기, 방 정보, 추노 노비 역할 랜덤

See merge request s08-webmobile1-sub2/S08P12D208!97

+++ chuno/src/main/java/com/leesfamily/chuno/room/model/RoomEntity.java
@@ -54,7 +54,7 @@ public class RoomEntity {
 //    private Point location;
 
     @Column(nullable = false)
-    private int radius;
+    private double radius;
 
     @Embedded
     private DateTime dateTime;
